Change to absolute coordinates and FFT func
Since all three bodies are described absolutely in terms of an inertial
coordinate system, the indentation should be calculated be only
considering the rotor position relative to the inner house, the same
applies to the alpha angle. Furthermore the FFT routine was moved to a
separate function, while plotting changes were made to the main
function.